requirements.txt
Pillow==8.1.0
pytesseract==0.3.7